CLN: Cleanup tests for .rank()
Work in Progress

1. Created new `pandas/tests/series/test_rank` and `pandas/tests/frame/test_rank`
2. Moved tests from `pandas/tests/test_stats` to `series` or `frame`
version of `test_rank`
3. move `rank` tests from `frame/test_analytics` to `frame/test_rank`
4. move `rank` tests from `series/test_analytics` to
`series/test_analytics`

See #15640
