+++ pandas/tests/frame/test_analytics.py
@@ -642,173 +642,6 @@ class TestDataFrameAnalytics(tm.TestCase, TestData):
         df.cumprod(0)
         df.cumprod(1)
 
-    def test_rank(self):
-        tm._skip_if_no_scipy()
-        from scipy.stats import rankdata
-
-        self.frame['A'][::2] = np.nan
-        self.frame['B'][::3] = np.nan
-        self.frame['C'][::4] = np.nan
-        self.frame['D'][::5] = np.nan
-
-        ranks0 = self.frame.rank()
-        ranks1 = self.frame.rank(1)
-        mask = np.isnan(self.frame.values)
-
-        fvals = self.frame.fillna(np.inf).values
-
-        exp0 = np.apply_along_axis(rankdata, 0, fvals)
-        exp0[mask] = np.nan
-
-        exp1 = np.apply_along_axis(rankdata, 1, fvals)
-        exp1[mask] = np.nan
-
-        tm.assert_almost_equal(ranks0.values, exp0)
-        tm.assert_almost_equal(ranks1.values, exp1)
-
-        # integers
-        df = DataFrame(np.random.randint(0, 5, size=40).reshape((10, 4)))
-
-        result = df.rank()
-        exp = df.astype(float).rank()
-        tm.assert_frame_equal(result, exp)
-
-        result = df.rank(1)
-        exp = df.astype(float).rank(1)
-        tm.assert_frame_equal(result, exp)
-
-    def test_rank2(self):
-        df = DataFrame([[1, 3, 2], [1, 2, 3]])
-        expected = DataFrame([[1.0, 3.0, 2.0], [1, 2, 3]]) / 3.0
-        result = df.rank(1, pct=True)
-        tm.assert_frame_equal(result, expected)
-
-        df = DataFrame([[1, 3, 2], [1, 2, 3]])
-        expected = df.rank(0) / 2.0
-        result = df.rank(0, pct=True)
-        tm.assert_frame_equal(result, expected)
-
-        df = DataFrame([['b', 'c', 'a'], ['a', 'c', 'b']])
-        expected = DataFrame([[2.0, 3.0, 1.0], [1, 3, 2]])
-        result = df.rank(1, numeric_only=False)
-        tm.assert_frame_equal(result, expected)
-
-        expected = DataFrame([[2.0, 1.5, 1.0], [1, 1.5, 2]])
-        result = df.rank(0, numeric_only=False)
-        tm.assert_frame_equal(result, expected)
-
-        df = DataFrame([['b', np.nan, 'a'], ['a', 'c', 'b']])
-        expected = DataFrame([[2.0, nan, 1.0], [1.0, 3.0, 2.0]])
-        result = df.rank(1, numeric_only=False)
-        tm.assert_frame_equal(result, expected)
-
-        expected = DataFrame([[2.0, nan, 1.0], [1.0, 1.0, 2.0]])
-        result = df.rank(0, numeric_only=False)
-        tm.assert_frame_equal(result, expected)
-
-        # f7u12, this does not work without extensive workaround
-        data = [[datetime(2001, 1, 5), nan, datetime(2001, 1, 2)],
-                [datetime(2000, 1, 2), datetime(2000, 1, 3),
-                 datetime(2000, 1, 1)]]
-        df = DataFrame(data)
-
-        # check the rank
-        expected = DataFrame([[2., nan, 1.],
-                              [2., 3., 1.]])
-        result = df.rank(1, numeric_only=False, ascending=True)
-        tm.assert_frame_equal(result, expected)
-
-        expected = DataFrame([[1., nan, 2.],
-                              [2., 1., 3.]])
-        result = df.rank(1, numeric_only=False, ascending=False)
-        tm.assert_frame_equal(result, expected)
-
-        # mixed-type frames
-        self.mixed_frame['datetime'] = datetime.now()
-        self.mixed_frame['timedelta'] = timedelta(days=1, seconds=1)
-
-        result = self.mixed_frame.rank(1)
-        expected = self.mixed_frame.rank(1, numeric_only=True)
-        tm.assert_frame_equal(result, expected)
-
-        df = DataFrame({"a": [1e-20, -5, 1e-20 + 1e-40, 10,
-                              1e60, 1e80, 1e-30]})
-        exp = DataFrame({"a": [3.5, 1., 3.5, 5., 6., 7., 2.]})
-        tm.assert_frame_equal(df.rank(), exp)
-
-    def test_rank_na_option(self):
-        tm._skip_if_no_scipy()
-        from scipy.stats import rankdata
-
-        self.frame['A'][::2] = np.nan
-        self.frame['B'][::3] = np.nan
-        self.frame['C'][::4] = np.nan
-        self.frame['D'][::5] = np.nan
-
-        # bottom
-        ranks0 = self.frame.rank(na_option='bottom')
-        ranks1 = self.frame.rank(1, na_option='bottom')
-
-        fvals = self.frame.fillna(np.inf).values
-
-        exp0 = np.apply_along_axis(rankdata, 0, fvals)
-        exp1 = np.apply_along_axis(rankdata, 1, fvals)
-
-        tm.assert_almost_equal(ranks0.values, exp0)
-        tm.assert_almost_equal(ranks1.values, exp1)
-
-        # top
-        ranks0 = self.frame.rank(na_option='top')
-        ranks1 = self.frame.rank(1, na_option='top')
-
-        fval0 = self.frame.fillna((self.frame.min() - 1).to_dict()).values
-        fval1 = self.frame.T
-        fval1 = fval1.fillna((fval1.min() - 1).to_dict()).T
-        fval1 = fval1.fillna(np.inf).values
-
-        exp0 = np.apply_along_axis(rankdata, 0, fval0)
-        exp1 = np.apply_along_axis(rankdata, 1, fval1)
-
-        tm.assert_almost_equal(ranks0.values, exp0)
-        tm.assert_almost_equal(ranks1.values, exp1)
-
-        # descending
-
-        # bottom
-        ranks0 = self.frame.rank(na_option='top', ascending=False)
-        ranks1 = self.frame.rank(1, na_option='top', ascending=False)
-
-        fvals = self.frame.fillna(np.inf).values
-
-        exp0 = np.apply_along_axis(rankdata, 0, -fvals)
-        exp1 = np.apply_along_axis(rankdata, 1, -fvals)
-
-        tm.assert_almost_equal(ranks0.values, exp0)
-        tm.assert_almost_equal(ranks1.values, exp1)
-
-        # descending
-
-        # top
-        ranks0 = self.frame.rank(na_option='bottom', ascending=False)
-        ranks1 = self.frame.rank(1, na_option='bottom', ascending=False)
-
-        fval0 = self.frame.fillna((self.frame.min() - 1).to_dict()).values
-        fval1 = self.frame.T
-        fval1 = fval1.fillna((fval1.min() - 1).to_dict()).T
-        fval1 = fval1.fillna(np.inf).values
-
-        exp0 = np.apply_along_axis(rankdata, 0, -fval0)
-        exp1 = np.apply_along_axis(rankdata, 1, -fval1)
-
-        tm.assert_numpy_array_equal(ranks0.values, exp0)
-        tm.assert_numpy_array_equal(ranks1.values, exp1)
-
-    def test_rank_axis(self):
-        # check if using axes' names gives the same result
-        df = pd.DataFrame([[2, 1], [4, 3]])
-        tm.assert_frame_equal(df.rank(axis=0), df.rank(axis='index'))
-        tm.assert_frame_equal(df.rank(axis=1), df.rank(axis='columns'))
-
     def test_sem(self):
         alt = lambda x: np.std(x, ddof=1) / np.sqrt(len(x))
         self._check_stat_op('sem', alt)
