+++ pandas/tests/series/test_analytics.py
@@ -969,207 +969,6 @@ class TestSeriesAnalytics(TestData, tm.TestCase):
             sc.drop_duplicates(keep=False, inplace=True)
             assert_series_equal(sc, s[~expected])
 
-    def test_rank(self):
-        tm._skip_if_no_scipy()
-        from scipy.stats import rankdata
-
-        self.ts[::2] = np.nan
-        self.ts[:10][::3] = 4.
-
-        ranks = self.ts.rank()
-        oranks = self.ts.astype('O').rank()
-
-        assert_series_equal(ranks, oranks)
-
-        mask = np.isnan(self.ts)
-        filled = self.ts.fillna(np.inf)
-
-        # rankdata returns a ndarray
-        exp = Series(rankdata(filled), index=filled.index, name='ts')
-        exp[mask] = np.nan
-
-        tm.assert_series_equal(ranks, exp)
-
-        iseries = Series(np.arange(5).repeat(2))
-
-        iranks = iseries.rank()
-        exp = iseries.astype(float).rank()
-        assert_series_equal(iranks, exp)
-        iseries = Series(np.arange(5)) + 1.0
-        exp = iseries / 5.0
-        iranks = iseries.rank(pct=True)
-
-        assert_series_equal(iranks, exp)
-
-        iseries = Series(np.repeat(1, 100))
-        exp = Series(np.repeat(0.505, 100))
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        iseries[1] = np.nan
-        exp = Series(np.repeat(50.0 / 99.0, 100))
-        exp[1] = np.nan
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        iseries = Series(np.arange(5)) + 1.0
-        iseries[4] = np.nan
-        exp = iseries / 4.0
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        iseries = Series(np.repeat(np.nan, 100))
-        exp = iseries.copy()
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        iseries = Series(np.arange(5)) + 1
-        iseries[4] = np.nan
-        exp = iseries / 4.0
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        rng = date_range('1/1/1990', periods=5)
-        iseries = Series(np.arange(5), rng) + 1
-        iseries.iloc[4] = np.nan
-        exp = iseries / 4.0
-        iranks = iseries.rank(pct=True)
-        assert_series_equal(iranks, exp)
-
-        iseries = Series([1e-50, 1e-100, 1e-20, 1e-2, 1e-20 + 1e-30, 1e-1])
-        exp = Series([2, 1, 3, 5, 4, 6.0])
-        iranks = iseries.rank()
-        assert_series_equal(iranks, exp)
-
-        # GH 5968
-        iseries = Series(['3 day', '1 day 10m', '-2 day', pd.NaT],
-                         dtype='m8[ns]')
-        exp = Series([3, 2, 1, np.nan])
-        iranks = iseries.rank()
-        assert_series_equal(iranks, exp)
-
-        values = np.array(
-            [-50, -1, -1e-20, -1e-25, -1e-50, 0, 1e-40, 1e-20, 1e-10, 2, 40
-             ], dtype='float64')
-        random_order = np.random.permutation(len(values))
-        iseries = Series(values[random_order])
-        exp = Series(random_order + 1.0, dtype='float64')
-        iranks = iseries.rank()
-        assert_series_equal(iranks, exp)
-
-    def test_rank_categorical(self):
-        # GH issue #15420 rank incorrectly orders ordered categories
-
-        # Test ascending/descending ranking for ordered categoricals
-        exp = pd.Series([1., 2., 3., 4., 5., 6.])
-        exp_desc = pd.Series([6., 5., 4., 3., 2., 1.])
-        ordered = pd.Series(
-            ['first', 'second', 'third', 'fourth', 'fifth', 'sixth']
-        ).astype(
-            'category',
-            categories=['first', 'second', 'third',
-                        'fourth', 'fifth', 'sixth'],
-            ordered=True
-        )
-        assert_series_equal(ordered.rank(), exp)
-        assert_series_equal(ordered.rank(ascending=False), exp_desc)
-
-        # Unordered categoricals should be ranked as objects
-        unordered = pd.Series(
-            ['first', 'second', 'third', 'fourth', 'fifth', 'sixth'],
-        ).astype(
-            'category',
-            categories=['first', 'second', 'third',
-                        'fourth', 'fifth', 'sixth'],
-            ordered=False
-        )
-        exp_unordered = pd.Series([2., 4., 6., 3., 1., 5.])
-        res = unordered.rank()
-        assert_series_equal(res, exp_unordered)
-
-        unordered1 = pd.Series(
-            [1, 2, 3, 4, 5, 6],
-        ).astype(
-            'category',
-            categories=[1, 2, 3, 4, 5, 6],
-            ordered=False
-        )
-        exp_unordered1 = pd.Series([1., 2., 3., 4., 5., 6.])
-        res1 = unordered1.rank()
-        assert_series_equal(res1, exp_unordered1)
-
-        # Test na_option for rank data
-        na_ser = pd.Series(
-            ['first', 'second', 'third', 'fourth', 'fifth', 'sixth', np.NaN]
-        ).astype(
-            'category',
-            categories=[
-                'first', 'second', 'third', 'fourth',
-                'fifth', 'sixth', 'seventh'
-            ],
-            ordered=True
-        )
-
-        exp_top = pd.Series([2., 3., 4., 5., 6., 7., 1.])
-        exp_bot = pd.Series([1., 2., 3., 4., 5., 6., 7.])
-        exp_keep = pd.Series([1., 2., 3., 4., 5., 6., np.NaN])
-
-        assert_series_equal(na_ser.rank(na_option='top'), exp_top)
-        assert_series_equal(na_ser.rank(na_option='bottom'), exp_bot)
-        assert_series_equal(na_ser.rank(na_option='keep'), exp_keep)
-
-        # Test na_option for rank data with ascending False
-        exp_top = pd.Series([7., 6., 5., 4., 3., 2., 1.])
-        exp_bot = pd.Series([6., 5., 4., 3., 2., 1., 7.])
-        exp_keep = pd.Series([6., 5., 4., 3., 2., 1., np.NaN])
-
-        assert_series_equal(
-            na_ser.rank(na_option='top', ascending=False),
-            exp_top
-        )
-        assert_series_equal(
-            na_ser.rank(na_option='bottom', ascending=False),
-            exp_bot
-        )
-        assert_series_equal(
-            na_ser.rank(na_option='keep', ascending=False),
-            exp_keep
-        )
-
-        # Test with pct=True
-        na_ser = pd.Series(
-            ['first', 'second', 'third', 'fourth', np.NaN],
-        ).astype(
-            'category',
-            categories=['first', 'second', 'third', 'fourth'],
-            ordered=True
-        )
-        exp_top = pd.Series([0.4, 0.6, 0.8, 1., 0.2])
-        exp_bot = pd.Series([0.2, 0.4, 0.6, 0.8, 1.])
-        exp_keep = pd.Series([0.25, 0.5, 0.75, 1., np.NaN])
-
-        assert_series_equal(na_ser.rank(na_option='top', pct=True), exp_top)
-        assert_series_equal(na_ser.rank(na_option='bottom', pct=True), exp_bot)
-        assert_series_equal(na_ser.rank(na_option='keep', pct=True), exp_keep)
-
-    def test_rank_signature(self):
-        s = Series([0, 1])
-        s.rank(method='average')
-        self.assertRaises(ValueError, s.rank, 'average')
-
-    def test_rank_inf(self):
-        pytest.skip('DataFrame.rank does not currently rank '
-                    'np.inf and -np.inf properly')
-
-        values = np.array(
-            [-np.inf, -50, -1, -1e-20, -1e-25, -1e-50, 0, 1e-40, 1e-20, 1e-10,
-             2, 40, np.inf], dtype='float64')
-        random_order = np.random.permutation(len(values))
-        iseries = Series(values[random_order])
-        exp = Series(random_order + 1.0, dtype='float64')
-        iranks = iseries.rank()
-        assert_series_equal(iranks, exp)
-
     def test_clip(self):
         val = self.ts.median()
